server/requirements.txt
flask>=3.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
python-chess>=1.999
zstandard>=0.21.0
certifi>=2024.0.0
